first task add item button
The user can enter the name of the product in the input field and click the Add button. After pressing the add button (or the Enter key), the product should be added to the end of the list. The input field should become empty and the cursor remains in the input field.